Insert join filters into iceberg filters

+++ bodo/pandas/_util.cpp
@@ -985,7 +985,7 @@ JoinFilterColStats::collect_all() {
 
 duckdb::unique_ptr<duckdb::TableFilterSet> JoinFilterColStats::insert_filters(
     duckdb::unique_ptr<duckdb::TableFilterSet> filters,
-    const std::vector<int64_t> column_projection) {
+    const std::vector<int> column_projection) {
     for (const auto &[col_idx, min_max_vec] : this->collect_all()) {
         // Find the position of col_idx in selected_columns
         for (const auto &[min, max] : min_max_vec) {
